实体类移动到api模块

+++ hutu-upms/hutu-upms-api/src/main/java/com/hutu/upms/api/entity/Log.java
@@ -44,7 +44,7 @@ public class Log implements Serializable {
     /**
      * 消耗时间
      */
-	private Integer spendTime;
+	private Long spendTime;
     /**
      * 根路径
      */
